Move some benchmarking functions into chatterbot module
